+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Developer release (not public to pip)
 * Add optional atribute with_header into Column.to_bytes() method. Default is True
 * Add NativeReader.read_block() method. This method will return python values as tuple object
+* Improve BlockReader.skip() method
 
 ## 0.2.5.dev1